requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
Django==4.1.3
et-xmlfile==1.1.0
flake8==6.0.0
idna==3.4
image==1.5.33
instaloader==4.9.5
mccabe==0.7.0
numpy==1.23.4
openpyxl==3.1.2
Pillow==9.3.0
plyer==2.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pygame==2.1.2
python-decouple==3.8
pytube==12.1.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
termcolor==2.1.0
tzdata==2022.6
urllib3==1.26.12
